refator(INJI-172): [Kiruthika] move storage limit reached check function to storage class
Issue Link https://mosip.atlassian.net/browse/INJI-172
